Modularised code, display bug on MainTaskFrame remains #25
