+++ mobile-browser-based-version/src/components/Home.vue
@@ -70,13 +70,11 @@
 </template>
 
 <script>
-
 import { mapMutations } from 'vuex';
 
 export default {
-  name: 'home',
+  name: 'Home',
   methods: {
-    ...mapMutations(['setIndexedDB']),
     goToTaskList() {
       this.$emit('gototasks');
       this.$router.push({
@@ -84,12 +82,5 @@ export default {
       });
     },
   },
-  mounted() {
-    /**
-     * At the start of the application, use IndexedDB by default if it is
-     * available.
-     */
-    this.setIndexedDB(window.indexedDB);
-  }
 };
 </script>
