--- mobile-browser-based-version/src/components/Trophee.vue
@@ -1,9 +1,0 @@
-<template>
-  Under construction ...
-</template>
-
-<script>
-export default {
-  name: 'trophee',
-};
-</script>
