+++ mobile-browser-based-version/src/components/main_frames/MainDescriptionFrame.vue
@@ -18,7 +18,6 @@
 </template>
 <script>
 import DescriptionFrame from '../building_frames/DescriptionFrame';
-import { defineComponent } from 'vue';
 
 export default {
   name: 'MainDescriptionFrame',
