+++ mobile-browser-based-version/src/main.js
@@ -1,5 +1,5 @@
 import { createApp } from 'vue';
-import App from './App.vue';
+import App from './components/App.vue';
 import router from './router';
 import '@/assets/css/tailwind.css';
 import '@/assets/css/styles.css';
