+++ mobile-browser-based-version/src/router/index.js
@@ -1,11 +1,9 @@
 import { createRouter, createWebHashHistory } from 'vue-router';
-import { defineComponent } from 'vue';
 
 // Some page components
 import Home from '../components/Home';
 import TaskList from '../components/TaskList';
 import Information from '../components/Information';
-import Trophee from '../components/Trophee';
 import NotFound from '../components/NotFound';
 
 var routes = [
@@ -25,15 +23,10 @@ var routes = [
     component: Information,
   },
   {
-    path: '/trophee',
-    name: 'trophee',
-    component: Trophee,
-  },
-  {
     path: '/:pathMatch(.*)*',
     name: 'not-found',
-    component: NotFound
-  }
+    component: NotFound,
+  },
 ];
 
 const router = createRouter({
